Limit files included in gem to those tracked by git
In building v0.5 of the `scss-lint` gem, I had unknowingly included some
local files that were not tracked by git, resulting in some code being
included that shouldn't have been.

Although the original reason we included files based on glob matching
was because we didn't want to unnecessary tie our gem to git, this event
brings to light one of the advantages of only including files that are
tracked by git, and that is you have to have committed the file before
you can "push" it with the gem.

Since I personally have a bunch of untracked experimental files in my
scss-lint repository, I'd rather us include files in the gem based on
whether or not they have been "committed" (i.e. have been code reviewed)
than just based on whether or not they exist.

This change also bumps the version number to 0.5.1 so that the broken
version can be weeded out.

Change-Id: I862b30fc2a2f43e8f9396d856bff84a62caf0532
Reviewed-on: https://gerrit.causes.com/17792
Tested-by: Shane da Silva <shane@causes.com>
Reviewed-by: Aiden Scandella <aiden@causes.com>