Add specs for CLI class
Adds a spec suite for the currently existing functionality for the `CLI`
class.

This requires a bit of a refactor, as the previous implementation made
the `initialize` implicitly invoked the `run` method, which made certain
parts of testing difficult. The refactor simplifies this by treating the
`initialize` method as just that: a place to initialize everything the
`CLI` object needs to run, without actually running anything.

This will make testing new features activated via command-line switches
much easier going forward.

Change-Id: I84f50eb4bf9b1dfcb81c8c121c67660ab238e5da
Reviewed-on: https://gerrit.causes.com/19665
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>