Call to_sass rather than to_s on nodes
`to_sass` does the right thing; `to_s` is for internal use and fails for
certain cases (e.g. node is a Sass::Script::List containing a function
call).

Ideally the spec I added here would be `subject.count.should == 1`, but
it's 0 presently due to the way we compare strings rather than actually
evaluate the Sass.

As a future improvement, we should see if we can determine the result of
these function calls (or at least improve the regex). This commit simply
keeps scss-lint from crashing on this input.

Fixes #11.

Change-Id: I51df3a678786a002c42c2375c3090d514387a867
Reviewed-on: https://gerrit.causes.com/19896
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Aiden Scandella <aiden@causes.com>