Add support for interpolation to SingleLinePerSelectorLinter
Previously, the `SingleLinePerSelectorLinter` would choke on
interpolated strings. This has been fixed by implementing a more
comprehensive check that properly deals with Sass::Script::Nodes in a
rule's comma sequence.

In the process of doing so, support was also added for not reporting a
lint when a selector in a multi-line rule had a comment on the end of
it, e.g.

    .foo, // A comment
    .bar {
    }

This invalid reporting was fixed by normalizing the whitespace appearing
around newline characters in the rule, since there were extra whitespace
characters being left behind when the Sass parser automatically removed
the comments from the parse tree. This fixes #10.

Change-Id: I300ce8a8fc5672c86585f375ff861b9ed50bc1f3
Reviewed-on: https://gerrit.causes.com/22121
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>