Fix ShorthandLinter to work with Sass script
Previously, the shorthand checker would incorrectly report a lint for a
property declaration like the following:

    padding: (10px / 12px) (15px / 12px);

This was because the regex was simply pulling out "10px 12px 15px 12px"
and assuming that the repeated "12px" could be shortened. However, in
this case the lint is incorrect.

The solution is to check whether or not we're dealing with a
Sass::Script::String, or a Sass::Script::List, and handle the two cases
differently. The Sass::Script::String case works as before, but the
Sass::Script::List case needs to individually call `to_sass` on each of
the items in the list and compare them to each other, rather than
calling `to_sass` on the list itself. This way expressions like:

    padding: $content-padding $content-padding;

...or...

    padding: (10px / 12px) 5px (10px / 12px);

...are correctly marked as lints, as we're comparing whole expressions
in the list.

Fixes #12

Change-Id: I90e6cd655d968e82b5f51900ee0ad00d4faa6daa
Reviewed-on: https://gerrit.causes.com/23139
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>