DRY up specs with custom matcher
A vast majority of specs for scss-lint test the linters themselves,
which usually consist of checking whether or not a linter reported a
lint, and which line the lint was on. This involves so much duplicate
code that it was prime for DRYing up with a custom matcher.

This adds a `report_lint` custom matcher and updates all specs to use
that matcher, resulting in a lot of duplicated code being removed.

Closes #18

Change-Id: I5fa1713397f42a73a952d85b33b117fb990196bd
Reviewed-on: https://gerrit.causes.com/23172
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>