Fix NotImplementedError in DirectiveNode monkey patch
Commit 500f4965 monkey-patched the parse tree node hierarchy of the Sass
parser in order to support linting Sass script nodes. This introduced a
bug where `scss-lint` would crash on any SCSS containing a `@media`
directive (and likely other subclasses of `DirectiveNode`).

This was due to the fact that `DirectiveNode` subclasses may not define
an implementation for `value`, which is used in the monkey-patched
implementation of `children` to find `Sass::Script::Node`s to lint.

The solution was to wrap the call to `value` with a rescue block
catching the `NotImplementedError`, and ignoring it.

Fixed #19

Change-Id: I143a5f4c5a6d712d93fa9b209d662fef15e38664
Reviewed-on: https://gerrit.causes.com/23260
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>