Add NoZeroBeforeDecimalLinter
Add a linter that checks for unnecessary leading zeros in numbers with a
fractional component. In other words, prefer `.5` over `0.5`.

This is unfortunately not a complete implementation, as numbers inside
Sass Script will not be detected in this way due to the way the Sass
parser replaces the original string with the resultant
`Sass::Script::Literal`. For example, '.5' will become the Float `0.5`,
so we can't detect after the fact whether the value was written with an
unnecessary leading zero. Concretely, values inside function or mixin
arguments will not be linted.

Luckily, this implementation handles a significant number of cases.

Perhaps one day we will contribute a commit to the upstream `sass` gem
to be able to handle the general case.

Change-Id: I069a4bd3095ddd02a5c31d1f5dfbe5c1b7ce77fa
Reviewed-on: https://gerrit.causes.com/27812
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>