Fix HexFormat linter false positives
The `HexFormat` linter would erroneously report a lint for ID selectors
that could be interpreted as hexadecimal color codes, e.g.

    #aabbcc {
      ...
    }

This was due to the simplistic regex the linter applied to every line of
the Sass file, without taking the parse tree into account.

In order to take the parse tree into account, however, the `Color` class
and the `Lexer` class of the Sass engine had to be monkey patched to
support storing the original color string (it normally gets lost during
the lexing phase). With this patch in place, the `Color` class now
exposes an attribute called `original` which allows any linter to
inspect the original string.

Fixes #36

Change-Id: Ie653295684cd79bfdaf94a10182af57d0bc84943
Reviewed-on: https://gerrit.causes.com/28953
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>