Add SpaceAfterComma linter
Add a linter which checks for the presence of a single space character
after a comma in argument lists or literal lists.

This lint is ignored when the comma is followed by a newline (i.e. the
arguments are split up over multiple lines).

This required a slight modification of the `character_at` helper to
return a newline character when the offset passed in as the second
argument results in the position being at the end of the line (i.e.
where the newline character would be).

Closes #26

Change-Id: I57db24658062fe8d6b85e4c9a1e754e4a83cb723
Reviewed-on: https://gerrit.causes.com/30681
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>