Fix Shorthand linter crash on numeric properties
The Sass parser appears to treat the following pieces of code
differently:

  p { margin: 5px; }

  p { margin: 5px }

The first results in the number being interpreted as a `Value::String`,
and the second results in the number being interpreted as a
`Value::Number`.

This discrepancy was causing the `Shorthand` linter to fail on the
second example.

Fixes #45

Change-Id: I06a372e95f18196b696ed9f82ea8cec581eb5267
Reviewed-on: https://gerrit.causes.com/30844
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>