Make indentation width configurable
Add support for generic configuration of linters and implement the first
use of this configuration system by making the indentation width of the
`Indentation` linter configurable.

In the process, augment the specs to always load the default
configuration for a linter from `scss-lint`'s' `config/default.yml`
file. This file should be seen as storing the canonical defaults for all
linters. However, for specs that wish to test different behavior based
on a linter's configuration, all linter specs can specify an explicit
configuration by declaring a `let` for `linter_config`.

Change-Id: I76e7114a9a156bd4ddd5b87391a6eef04648eebd
Reviewed-on: https://gerrit.causes.com/30923
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>