Switch to semantic exit codes for CLI#halt
`CLI#halt` was being invoked in several places with an exit code of -1.
Unix exit codes are traditionally limited to integer args in the range
0-255 [1][2], and out of range exit values result in unexpected
behavior. This commit reverts the error codes to use the catchall error
exit code 1, although if one is feeling fancy an exit of 65 (EX_DATAERR -
input data incorrect) may be valid in some places.

[1] - http://en.wikipedia.org/wiki/Exit_status
[2] - http://tldp.org/LDP/abs/html/exitcodes.html

Change-Id: Ic5099a7c3ade5f6a5475ebd3998150526354e26b
Reviewed-on: https://gerrit.causes.com/31574
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>