Improve regex for ColorKeyword linter
When linting string identifiers returned by the Sass parser, the
`ColorKeyword` linter would incorrectly report lints for words like
`item-blue`, since it contains a substring `blue` which is a color
keyword.

The solution was to improve the tokenizer regex by ensuring a
color-keyword is surrounded by start-of-line/end-of-line or whitespace.

Fixes #52

Change-Id: Ib3a60a7d80146c3bcb2ada0677d3d3cd261bed82
Reviewed-on: https://gerrit.causes.com/31757
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>