Add -f/--format command-line flag
Commit 09d90bf added support for changing the output format of
`scss-lint` to return XML.

Adding top-level command-line flags can lead to future problems when we
start overloading the number of flags provided by the tool. Instead,
convert the `--xml` flag to use a more generic `--format` flag. This has
the added benefit making it easier in the future to add new formatters
that are automatically recognized by the tool.

Change-Id: I30180fc39789413a6b72438d359adae5d5589a8e
Reviewed-on: https://gerrit.causes.com/31883
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>