Fix error when using `@keyframe`s
When using a `@keyframe`, the SelectorDepth linter would run into the
following error:

    Failure/Error: Unable to find matching line from backtrace
    NoMethodError:
      undefined method `rest' for ["0%", ""]:Array

This happens because the Sass Selector Sequence members may be,
SimpleSequence objects, Strings, or Arrays[1], but some code in the
SelectorDept linter was written with the assumption that the members may
be only either SimpleSequence objects or Strings but not Arrays.

Unfortunately, the documentation isn't very clear on when these members
may be Arrays, but it appears to happen for `@keyframe` percentages.
Since this is the first time we have run into this problem, we are
simply skipping over these Arrays because selector depth does not apply
when in a `@keyframe`.

Fixes #74

[1]: http://sass-lang.com/documentation/Sass/Selector/Sequence.html#members-instance_method

Change-Id: I549677d179335fddebb395085d70dea75b897130
Reviewed-on: https://gerrit.causes.com/33038
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>