Partially fix false positives in SpaceBetweenParens
The `SpaceBetweenParens` linter would report a lint for invalid
parentheses spacing in comments.

This is actually a hard problem to solve, because the Sass parser
doesn't report valid source ranges for certain node types or
expressions.

One such example is:

    p {
      margin: ( value);
    }

In the above example, the source range for the literal value in brackets
is " value)". In other words, we have no way of knowing where the
opening parenthesis is if this expression were to be spread over
multiple lines, and robustly detecting this case is incredibly involved.

For the nature of this check, the juice simply isn't worth the squeeze
unless the Sass parser gets more accurate with its source range
reporting.

As a compromise, add some handling that will deal with the majority of
cases by ignoring parentheses after comment line openers (`//` or `/*`).
It will still miss some cases, but it's good enough for free software.

Change-Id: I8d35255236e7fda598beff0c6a06bae235fcfb22
Reviewed-on: http://gerrit.causes.com/36421
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>