Fix TrailingSemicolonAfterPropertyValue multiline bug
The `TrailingSemicolonAfterPropertyValue` linter would incorrectly
report a lint for properties split up across multiple lines, e.g.

    .widget {
      background:
        top repeat-x url("top-image.jpg"),
        right repeat-y url("right-image.jpg"),
        bottom repeat-x url("bottom-image.jpg"),
        left repeat-y url("left-image.jpg");
    }

This was due to a special case in `Linter#source_from_range` which no
longer seems to be needed. Remove it to solve the problem.

The `TrailingSemicolonAfterPropertyValue` linter itself was also changed
to always report the last line of the property instead of the first.

Fixes #106

Change-Id: Ib44f841cc46d9ab6cd76819e21e551747ef6c42e
Reviewed-on: http://gerrit.causes.com/36889
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>