Fix source_from_range with lines with no semicolon
The `StringQuotes` linter would fail on `@import` lines with no
semicolon (which is allowed if they are the last line in the file).

This was causes by the fact that the Sass parser will return the end
position line after the actual last line if there is no semicolon.

Fix by returning an empty string if attempting to fetch the contents of
a nonexistent line.

Change-Id: I5a75a22615a3131dc1571bc1768df02d4f3a3912
Reviewed-on: http://gerrit.causes.com/36952
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>