Inspect parse tree directly in SpaceBeforeBrace
The `SpaceBeforeBrace` linter suffered from a class of bugs where a
false positive would be reported for braces that weren't actually
opening braces (e.g. braces in comments).

This was due to the naive regex implementation that did not take the
parse tree into account.

Fix this problem by inspecting the parse tree directly.

Change-Id: I1e10970553a30fa2d0fb0bd73ad802a1a8c02806
Reviewed-on: http://gerrit.causes.com/37227
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>