Add support for additional styles to SpaceAfterPropertyColon
Add support for specifying that property values should not have any
spaces separating them from the colon, or that the property values
should all be aligned, via the `style` option on the
`SpaceAfterPropertyColon` linter (which replaces the
`allow_extra_spaces` option).

The main purpose was to add support for alignment, but while we were
here we decided to add the `no_space` option since it was easy to do.

Change-Id: I9c71c6c0b593ddf59d0f286e7f8f08ee09384909
Reviewed-on: http://gerrit.causes.com/38492
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>