Upgrade to RSpec 3
Bring us up to date with the latest and greatest in spec writing
technology.

The only quirk with this update was strange behavior coming from the
use of `any_instance` in the `Runner` spec. This looked to be a bug with
RSpec (since we use `any_instance` without problem elsewhere), so the
workaround was to add a method whose only purpose is to be stubbed. It's
not great, but it's isolated for now.

Change-Id: I12644e387b518946fc167c2466b14ccf758e106f
Reviewed-on: http://gerrit.causes.com/40447
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>