Add detection of mixture of tabs/spaces
Previously, the `Indentation` linter would not report an error when tabs
were mixed with spaces, so if a line started with the right number of
appropriate characters (e.g. 1 tab), it could be followed by any number
of spaces and still pass the check.

Fix this problem by checking for a mixture of tabs and spaces and
reporting the problem.

Change-Id: I9503eafa40333887014f0649163f43a0f7bc6263
Reviewed-on: http://gerrit.causes.com/40781
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>