Add ability to configure linter severity
Add support for linters to respect a new `severity` option, which
dictates whether a lint reported by a linter is an error or a warning.

At this point in time this only changes the output of `scss-lint`, not
its exit code (i.e. the presence of warnings will still result in a
non-zero exit status identical to the presence of errors).

A follow-up commit will change `scss-lint` to return different error
codes depending on whether warnings or errors are present.

Change-Id: Ia322db6300ac162dd760c1342803019004ef0b4e
Reviewed-on: http://gerrit.causes.com/40805
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>