Add ability to specify preset sort order
Add support to the `PropertySortOrder` linter to take a string for the
`order` config option, which then attempts to load a pre-defined
ordering from the `data/property-sort-orders` directory.

This is intended to provide some out-of-the-box orderings to help
developers not have to come up with their own sort order.

At the moment, the only sort order supported is `concentric`
(http://rhodesmill.org/brandon/2011/concentric-css/).

Change-Id: I2bf1d09a3edcc8d776d0f21f49eb95fb80f4c115
Reviewed-on: http://gerrit.causes.com/40836
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>